Tidy up tests for 43Ca+.
There was previously a fixme here about the values we calculare
not matching the ones from the Benhelm paper. I tracked this
difference down to the fact that they were using a measured gJ
for the D5/2 level while we were using the Lande g factor. NB
they are also using a different (older) value of gI to this
toolkit, although that does not change the calculations meaningfully.
